Rearranged server files in order to accommodate testing
