+++ __tests__/server-test.js
@@ -1,0 +1,15 @@
+// create a new review query
+// send that query to the server
+// should get a response 201
+
+// beforeAll(() => {
+//   return initializeDatabase();
+// });
+
+// test('server updates database on POST request', () => {
+//   //
+// })
+
+test('it adds two numbers', () => {
+  expect(1 + 1).toBe(2);
+});
